==> analysis/01A_clean-outcome-cov.Rmd <==
---
title: "01A_clean-outcome-cov"
author: "Mia Navarro"
date: "2024-05-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Project Overview

Clean up the following variables:

*  `Y`: visits by CSPS and month (count)
*  `A`: violence by year in Nouna (count, number of events) SEE 01.2.1 FILE
*  `C1`: season (wet/dry)
*  `C2`: Ramadan (binary)

### 0. Configure environment. Load data.
```{r config, show=F, warning=F, message=F}

source(here::here("0-config.R"))

vill_codes_names <- read_csv(paste0(raw_data_path,"vill_codes_names.csv"))

vill_csps_gps <- read_csv(paste0(raw_data_path,"vill_csps_gps.csv"), 
                          col_types = cols(csps = col_integer(), 
                                           village = col_integer()))

passive <- read_csv(paste0(raw_data_path,"passive (1).csv"))

census <- read_csv(paste0(raw_data_path,"CHAT_child_census_dataset.csv"))

```


### 1. Generate df with village IDs, names, population sizes.
```{r village data, warning=F, message=F}

vill_census <- 
  census %>% 
  distinct(childID, .keep_all = T) %>%
  group_by(village) %>%
  summarise(n_census=n())#,
            # n_06=n(ageInMonths %in% 1:6),
            # n_712=n(ageInMonths %in% 7:12),
            # n_1318=n(ageInMonths %in% 13:18),
            # n_1924=n(ageInMonths %in% 19:24),
            # n_2536=n(ageInMonths %in% 25:36),
            # n_3748=n(ageInMonths %in% 37:48),
            # n_4960=n(ageInMonths %in% 49:60)) 
  
vill_name_pop <- 
  vill_census %>%
  left_join(vill_codes_names, by="village") %>%
  arrange(village)

vill_name_pop #n=249 of 288 villages had census data.

write.csv(vill_name_pop, paste0(temp_data_path,"vill_name_pop.csv"))

```

### 2. Clean wide data,
```{r wide data, warning=F, message=F}

df_2 <- 
  passive %>%
  mutate(visit_date = as.Date(starttime),
         visit_year = year(starttime),
         visit_month = month(starttime),
         dx_type = case_when(diagnosis__57 == 1 ~ 1, #sev malaria
                             diagnosis__58 == 1 ~ 1, #non-sev malaria
                             diagnosis__39 == 1 ~ 2, #broncho-pneum
                             diagnosis__40 == 1 ~ 2, #low pneum
                             diagnosis__64 == 1 ~ 2, #pneum
                             diagnosis__21 == 1 ~ 3, #diarrhea
                             .default = 0)) %>%
  left_join(vill_name_pop, by=join_by(village==village)) %>% #add village info
  mutate(inv_wt_vill_pop = 1/n_census) %>% #visit weight by village pop size
  dplyr::filter(!grepl("Outside Nouna",csps_name)) %>% 
  dplyr::filter(!is.na(csps_name))

df_3 <- 
  df_2  %>%
  group_by(csps,visit_year,visit_month) %>%
  summarise(n_visit = n()) # %>% dplyr::filter(!csps %in% c(28,50)) #%>% arrange(desc(n_visit_per100))

df_3_dx <- 
  df_2 %>% 
  group_by(csps,visit_year,visit_month,dx_type) %>%
  summarise(n_visit = n()) # %>% dplyr::filter(!csps %in% c(28,50)) #%>% arrange(desc(n_visit_per100))

  
length(unique(df_3$village)) #n=257 (7 more than in census df)
length(unique(vill_csps_gps$villcode)) #n=226 (31 fewer than in passive surveillance df)

head(df_3); head(df_3_dx)
```

### 3. Create subsets of data.
```{r download, warning=F, message=F}

df_all_v1 <- df_3

df_mal_v1 <- df_3_dx %>% filter(dx_type == 1)
df_pne_v1 <- df_3_dx %>% filter(dx_type == 2)
df_dia_v1 <- df_3_dx %>% filter(dx_type == 3)

#Tranpose; add NA values for unobserved months.
df_all_v2 <- merge_long_data(df_all_v1,4:5)
df_mal_v2 <- merge_long_data(df_mal_v1,4:5) 
df_pne_v2 <- merge_long_data(df_pne_v1,4:5)
df_dia_v2 <- merge_long_data(df_dia_v1,4:5)

```

### 5. Create + add offset variable, csps population size
```{r offset-var}

# calculate how many children from each village actually visited a CSPS
csps_vill_n_actual <- 
  passive %>% 
  group_by(csps, village) %>% 
  filter() %>% 
  summarise(n_visit_act_csps_vill=n()) %>% 
  filter(csps>0) %>% 
  arrange(csps,village)


csps_vill_n <- 
  csps_vill_n_actual %>% 
  left_join(vill_name_pop, by=join_by(village==village)) %>% # merge vill pop size
  group_by(csps) 

csps_vill_n_expected <-
  csps_vill_n %>%
  summarise(n_visit_exp_max_csps=sum(n_census, na.rm=T)) # calculate how many children a CSPS could have served

csps_n_tot <- 
  passive %>% 
  group_by(csps) %>%
  summarise(offset_n_visit_tot_csps=n()) %>% 
  filter(csps>0) %>% 
  arrange(csps)

df_all_v3 <- merge_offset_percap(df_all_v2)
df_mal_v3 <- merge_offset_percap(df_mal_v2)
df_pne_v3 <- merge_offset_percap(df_pne_v2)
df_dia_v3 <- merge_offset_percap(df_dia_v2)

```


### 6. Save data.
```{r save-data}

write_csv(df_all_v3, paste0(temp_data_path,"df_all_temp_",Sys.Date(),".csv"))
write_csv(df_mal_v3, paste0(temp_data_path,"df_mal_temp_",Sys.Date(),".csv"))
write_csv(df_pne_v3, paste0(temp_data_path,"df_pne_temp_",Sys.Date(),".csv"))
write_csv(df_dia_v3, paste0(temp_data_path,"df_dia_temp_",Sys.Date(),".csv"))

```
