==> analysis/01B_clean-exposure.Rmd <==
---
title: "01B_clean-exposure"
author: "Mia Navarro"
date: "2024-05-18"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Project Overview

Clean up the following variables:

*  `Y`: visits by CSPS and month (count) SEE 01.2.0 FILE
*  `A`: violence by year in Nouna (count, number of events)
*  `C1`: season (wet/dry) SEE 01.2.0 FILE
*  `C2`: Ramadan (binary) SEE 01.2.0 FILE

### 0. Configure environment. Load data.
```{r config, show=F, warning=F, message=F}

source(here::here("0-config.R"))

vill_codes_names <- read_csv(paste0(raw_data_path,"vill_codes_names.csv"))

vill_csps_gps <- read_csv(paste0(raw_data_path,"vill_csps_gps.csv"), 
                          col_types = cols(csps = col_integer(), 
                                           village = col_integer()))

csps_codes <- read_csv(paste0(raw_data_path,"csps_codes.csv"))

# Extract CSPS GPS data (`csps_lat` and `csps_lon`).
csps_gps <- vill_csps_gps %>%
  dplyr::select(c(csps,csps_lat_new,csps_lon_new)) %>%
  filter(!duplicated(csps)) %>%
  rename(csps_lat = csps_lat_new,
         csps_lon = csps_lon_new) %>%
  arrange(csps)

write_csv(csps_gps, paste0(temp_data_path, "csps_gps.csv"))

acled_data <- read_csv(paste0(raw_data_path,"ACLED_2020-02-01-2023-04-01-Burkina_Faso v2.csv"))

#temp data from 01A file
date_data <- "2024-09-26"

df_all_temp <- read_csv(paste0(temp_data_path,"df_all_temp_",date_data,".csv"))
df_mal_temp <- read_csv(paste0(temp_data_path,"df_mal_temp_",date_data,".csv"))
df_pne_temp <- read_csv(paste0(temp_data_path,"df_pne_temp_",date_data,".csv"))
df_dia_temp <- read_csv(paste0(temp_data_path,"df_dia_temp_",date_data,".csv"))

```

### Clean violence data.
```{r viol, warning=F, message=F}

df_violence <- 
  acled_data %>%
  mutate(date = as.Date(event_date, format = "%d %B %Y")) %>%
  mutate(month = month(date)) %>%
  relocate(date, .after=event_date) %>%
  relocate(month, .after=year) %>%
  arrange(date) %>%
  filter(admin2 %in% c("Kossi")) %>%
  dplyr::select(c(year,month,longitude,latitude))


df_violence$id_event <- 1:nrow(df_violence)

df_violence <-
  df_violence %>%
  relocate(id_event, .before=year)

df_violence

write_rds(df_violence, paste0(temp_data_path,"df_violence.rds"))

acled_data %>% group_by(disorder_type) %>% summarise(n=n())
acled_data %>% group_by(event_type) %>% summarise(n=n())

```

### Determine max distance to consider a CSPS "close" to violent event.
```{r distance, warning=F, message=F}

mean(vill_csps_gps$dist_csps); sd(vill_csps_gps$dist_csps)
median(vill_csps_gps$dist_csps); quantile(vill_csps_gps$dist_csps, probs=seq(0,1,0.1))

hist(vill_csps_gps$dist_csps, breaks = 14)

maxdist_km = 20 #We'll use 20 km.

```

### Loop through CSPS. Append row to violence data for each CSPS with flags if within 20 km.
```{r, warning=F, message=F}

# calculate distances of each csps from each event in meters
for (hf in csps_gps$csps){
  
  near_viol <- 
    recipe(~., data = df_violence) %>%
    update_role(id_event, new_role = "location") %>%
    step_geodist(
      lat = latitude, 
      lon = longitude, 
      log = FALSE,
      ref_lat = csps_gps$csps_lat[csps_gps$csps==hf], 
      ref_lon = csps_gps$csps_lon[csps_gps$csps==hf],
      is_lat_lon = TRUE
      ) %>%
    prep(training = df_violence)
  
  df_violence <- bake(near_viol, new_data = NULL)
  colnames(df_violence)[ncol(df_violence)] <- paste0("dist_csps_",hf)
  
}

df_viol2 <- df_violence #make copy
df_viol2[,6:50] <- df_viol2[,6:50]/1e3 #convert m to km


# define `near_viol` == 1 if within 20 km, == 2 if beyond 20 km
for (hf in csps_gps$csps){
  
  tempdf <- 
    df_viol2 %>% 
    group_by(year,month) %>% 
    summarise(
      near_violence = 
        ifelse(min(get(paste0("dist_csps_",hf)))<=maxdist_km,2,1))
  
  df_viol2 <- 
    df_viol2 %>% 
    left_join(tempdf, by = c("year", "month"))
  
}

colnames(df_viol2)[51:95] <- csps_gps$csps
df_viol2$date <- as.Date(paste0(df_viol2$month,"/01/",df_viol2$year), "%m/%d/%Y")

df_viol2 <-
  df_viol2 %>%
  relocate(date, .before=year)

df_viol2

```

### Create two dfs: one for categorial (`near_violence`) and one for continuous (`min_dist_to_viol`)
```{r}
# categorial
df_viol3_cat <- df_viol2 %>% dplyr::select(c(2,52:96)) 

df_viol4_cat <- 
  make_viol_df_long(df_viol3_cat) %>% 
  arrange(csps, year, month, -near_violence) %>% 
  distinct()

# continuous
df_viol3_cont <- df_viol2 %>% dplyr::select(c(2,7:51))

colnames(df_viol3_cont)[2:46] <- csps_codes$csps

df_viol4_cont <- 
  make_viol_df_long(df_viol3_cont, "dist_to_viol") %>%
  group_by(csps, year, month) %>%
  summarise(min_dist_to_viol = min(dist_to_viol, na.rm=T)) %>%
  arrange(csps, year, month, min_dist_to_viol)
```


### 2C: Merge data with outcome and covariate data frames. Download.
```{r merge, warning=F, message=F}

df_all_v4 <- generate_final_data(df_all_temp)
df_mal_v4 <- generate_final_data(df_mal_temp)
df_pne_v4 <- generate_final_data(df_pne_temp)
df_dia_v4 <- generate_final_data(df_dia_temp)

head(df_all_v4)

write_csv(df_all_v4, paste0(final_data_path,"df_all_",Sys.Date(),".csv"))
write_csv(df_mal_v4, paste0(final_data_path,"df_mal_",Sys.Date(),".csv"))
write_csv(df_pne_v4, paste0(final_data_path,"df_pne_",Sys.Date(),".csv"))
write_csv(df_dia_v4, paste0(final_data_path,"df_dia_",Sys.Date(),".csv"))

write_rds(df_all_v4, paste0(final_data_path,"df_all_",Sys.Date(),".rds"))
write_rds(df_mal_v4, paste0(final_data_path,"df_mal_",Sys.Date(),".rds"))
write_rds(df_pne_v4, paste0(final_data_path,"df_pne_",Sys.Date(),".rds"))
write_rds(df_dia_v4, paste0(final_data_path,"df_dia_",Sys.Date(),".rds"))

```
